% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCG.Levels.R
\name{BCG.Levels}
\alias{BCG.Levels}
\title{BCG Levels}
\usage{
BCG.Levels(df.membership, df.rules)
}
\arguments{
\item{df.membership}{Data frame of metric memberships (long format, the same as the output of BCG.Membership).}

\item{df.rules}{Data frame of BCG model rules.}
}
\value{
Returns a data frame of results in the wide format.
}
\description{
Biological Condition Gradient Level assignment given metric memberships.
}
\details{
Input is metric memberships and a rules tables.  Output are
}
\examples{
library(readxl)

# Calculate Metrics
df.samps.bugs <- read_excel(system.file("./extdata/Data_BCG_PacNW.xlsx"
                                        , package="BCGcalc"))
myDF <- df.samps.bugs
df.metric.values.bugs <- metric.values(myDF, "bugs")

# Import Rules
df.rules <- read_excel(system.file("./extdata/Rules.xlsx"
                             , package="BCGcalc"), sheet="BCG_PacNW_2018") 

# Calculate Membership
df.Membership <- BCG.Membership(df.metric.values.bugs, df.rules)

# Calculate Levels
df.Levels <- BCG.Levels(df.Membership, df.rules)

# show results
View(df.Levels)

}
