% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCG.Membership.R
\name{BCG.Membership}
\alias{BCG.Membership}
\title{BCG Membership}
\usage{
BCG.Membership(df.metrics, df.rules, input.shape = "wide")
}
\arguments{
\item{df.metrics}{Wide data frame with metric values to be evaluated.}

\item{df.rules}{Data frame of metric thresholds to check.}

\item{input.shape}{Shape of df.metrics; wide or long.  Default is wide.}
}
\value{
Returns a data frame of results in the long format.
}
\description{
Biological Condition Gradient fuzzy membership for metrics.
}
\details{
used reshape2 package.....
}
\examples{
library(readxl)

# Calculate Metrics
df.samps.bugs <- read_excel("./data-raw/Data_BCG_PacNW.xlsx")
myDF <- df.samps.bugs
df.metric.values.bugs <- metric.values(myDF, "bugs")

# Import Rules
df.rules.PacNW <- read_excel("./inst/extdata/Rules.xlsx", sheet="BCG_PacNW_2018") 

# Run function
df.Membership <- BCG.Membership(df.metric.values.bugs, df.rules.PacNW)

# show results
View(df.Membership)

}
