% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCG.Tiers.R
\name{BCG.Tiers}
\alias{BCG.Tiers}
\title{BCG Tiers}
\usage{
BCG.Tiers(df.levels)
}
\arguments{
\item{df.levels}{Wide data frame with level memberships (0-1).  L1 to L6.}
}
\value{
Returns a data frame of results in the wide format.
}
\description{
Biological Condition Gradient tier assignment given Level memberships.
}
\details{
Input is L1 to L6 with membership values of 0 to 1.  
Result is 1st and 2nd Tiers (narrative and membership).  
Also give close.
}
\examples{

# construct a dummy dataset
L1 <- rep(0, 10)
L2 <- c(0.4, 0, 0.4, rep(0,7))
L3 <- c(0.6, 0, 0.6, 0, 0.42, 0, 1, 1, 0.22, 0.33)
L4 <- c(0, 0.9, 0, 0, 0.58, 0.05, 0, 0, 0.78, 0.67)
L5 <- c(0, 0.1, 0, 1, 0, 0.95, rep(0,4))
L6 <- rep(0, 10)
SampID <- LETTERS[1:10]
df.levels <- as.data.frame(SampID, stringsAsFactors=FALSE)
df.levels[,"L1"] <- L1
df.levels[,"L2"] <- L2
df.levels[,"L3"] <- L3
df.levels[,"L4"] <- L4
df.levels[,"L5"] <- L5
df.levels[,"L6"] <- L6

df.Tiers <- BCG.Tiers(df.levels)

View(df.Tiers)

}
