% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCGcalc.R
\docType{package}
\name{BCGcalc}
\alias{BCGcalc}
\alias{BCGcalc-package}
\title{BCGcalc: MEtric calculation and other tasks related to BCG}
\description{
Enables users to employ various tools and calculations related to BCG.
}
\section{BCGcal functions}{

metric.values
qc.checks
rarify
...
}

