% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_bio2rarify}
\alias{data_bio2rarify}
\title{rarify example data}
\format{A data frame with 74 rows and 3 variables:
\describe{
   \item{SampRep}{Station code}
   \item{tax}{}
   \item{count}{}
}}
\source{
example data
}
\usage{
data_bio2rarify
}
\description{
A dataset with example benthic macroinvertebrate data to be used with the rarify function.
}
\keyword{datasets}
