% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_GIS_BCGclass}
\alias{data_GIS_BCGclass}
\title{ORWA BCG classes}
\format{
A GIS data frame with 3 observations on the following 2 variables.:
\describe{
   \item{BCGclass_v}{a character vector}
   \item{geometry}{a sfc_MULTIPOLYGON}
}
}
\usage{
data_GIS_BCGclass
}
\description{
GIS data for BCG classification for Oregon-Washington
}
\keyword{datasets}
