% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_GIS_eco3_orwa}
\alias{data_GIS_eco3_orwa}
\title{Level 3 Ecoregions, ORWA}
\format{
A GIS data frame with 23 observations on the following 8 variables.:
\describe{
  \item{\code{AREA}}{a numeric vector}
  \item{\code{PERIMETER}}{a numeric vector}
  \item{\code{USECO_3_}}{a numeric vector}
  \item{\code{USECO_3_ID}}{a numeric vector}
  \item{\code{ECO}}{a character vector}
  \item{\code{LEVEL3}}{a numeric vector}
  \item{\code{LEVEL3_NAM}}{a character vector}
  \item{\code{geometry}}{a sfc_MULTIPOLYGON}
}
}
\usage{
data_GIS_eco3_orwa
}
\description{
GIS data with Level 3 ecoregions for Oregon and Washington
}
\keyword{datasets}
