% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_GIS_eco3_orwa_bcg}
\alias{data_GIS_eco3_orwa_bcg}
\title{ORWA Eco3 BCG}
\format{
A data frame with 12 observations on the following 6 variables.:
\describe{
  \item{\code{US_L3CODE}}{a character vector}
  \item{\code{US_L3NAME}}{a character vector}
  \item{\code{BCG}}{a character vector}
  \item{\code{Fill}}{a character vector}
  \item{\code{BCG_Valid}}{a character vector}
  \item{\code{geometry}}{a sfc_MULTIPOLYGON}
}
}
\usage{
data_GIS_eco3_orwa_bcg
}
\description{
ORWA Level 3 ecoregions and BCG classes for Oregon and Washington
}
\keyword{datasets}
