% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metrics_scoring}
\alias{metrics_scoring}
\title{Parameter List}
\format{A data frame with 69 rows and 8 variables:
\describe{
   \item{parm}{}
   \item{parmSource}{}
   \item{parmCat}{}
   \item{parmName}{}
   \item{parmCalc}{}
   \item{parmNamelc}{}
   \item{parmUnits}{}
   \item{parmSource}{}
   \item{parmRecensor}{}
   \item{parmRO1}{}
   \item{parmRO2}{}
   \item{parmTrend}{}
   \item{logTrans}{}
   \item{order}{}
   \item{trendIncrease}{}
}}
\usage{
metrics_scoring
}
\description{
Long Description
}
\keyword{datasets}
