% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.checks.R
\name{qc.checks}
\alias{qc.checks}
\title{QC checks on metric values}
\usage{
qc.checks(df.metrics, df.checks, input.shape = "wide")
}
\arguments{
\item{df.metrics}{Wide data frame with metric values to be evaluated.}

\item{df.checks}{Data frame of metric thresholds to check.}

\item{input.shape}{Shape of df.metrics; wide or long.  Default is wide.}
}
\value{
Returns a data frame of SampleID checks and results; Pass and Fail.
}
\description{
Long Description
}
\details{
used reshape2 package
}
\examples{
library(readxl)

# Calculate Metrics
df.samps.bugs <- read_excel(system.file("./extdata/Data_BCG_PacNW.xlsx"
                                        , package="BCGcalc"))
myDF <- df.samps.bugs
df.metric.values.bugs <- metric.values(myDF, "bugs")

# Import Checks
df.checks <- read_excel(system.file("./extdata/MetricFlags.xlsx"
                                          , package="BCGcalc"), sheet="Flags") 

# Run Function
df.flags <- qc.checks(df.metric.values.bugs, df.checks)

# Summarize Results
table(df.flags[,"CheckName"], df.flags[,"Flag"])
}
