% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TaxaMaster_Ben_BCG_PacNW}
\alias{TaxaMaster_Ben_BCG_PacNW}
\title{USGS Gages}
\format{A data frame with 684 rows and 19 variables:
\describe{
   \item{TaxaID}{taxonomic identifier}
   \item{Phylum}{Phylum}
}}
\usage{
TaxaMaster_Ben_BCG_PacNW
}
\description{
Long Description
}
\keyword{datasets}
