% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCG.Level.Assignment.R
\name{BCG.Level.Assignment}
\alias{BCG.Level.Assignment}
\title{BCG Level Assignment}
\usage{
BCG.Level.Assignment(df.level.membership)
}
\arguments{
\item{df.level.membership}{Wide data frame with level memberships (0-1).  L1 to L6.}
}
\value{
Returns a data frame of results in the wide format.
}
\description{
Biological Condition Gradient level assignment (1st and 2nd) given Level memberships.
}
\details{
Input is L1 to L6 with membership values of 0 to 1.  
Result is 1st (primary) and 2nd (secondary) Level.  
Also give close.

Add QC Checks.
}
\examples{

# construct a dummy dataset
L1 <- rep(0, 10)
L2 <- c(0.4, 0, 0.4, rep(0,7))
L3 <- c(0.6, 0, 0.6, 0, 0.42, 0, 1, 1, 0.22, 0.33)
L4 <- c(0, 0.9, 0, 0, 0.58, 0.05, 0, 0, 0.78, 0.67)
L5 <- c(0, 0.1, 0, 1, 0, 0.95, rep(0,4))
L6 <- rep(0, 10)
SampID <- LETTERS[1:10]
df.Level.Membership <- as.data.frame(SampID, stringsAsFactors=FALSE)
df.Level.Membership[,"L1"] <- L1
df.Level.Membership[,"L2"] <- L2
df.Level.Membership[,"L3"] <- L3
df.Level.Membership[,"L4"] <- L4
df.Level.Membership[,"L5"] <- L5
df.Level.Membership[,"L6"] <- L6

# Run Function
df.Levels <- BCG.Level.Assignment(df.Level.Membership)

# Show Results
View(df.Levels)

# Save Results
write.table(df.Levels, "Levels.tsv"
            , row.names=FALSE, col.names=TRUE, sep="\\t")

}
