% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCG.Level.Membership.R
\name{BCG.Level.Membership}
\alias{BCG.Level.Membership}
\title{BCG Level Membership}
\usage{
BCG.Level.Membership(df.metric.membership, df.rules)
}
\arguments{
\item{df.metric.membership}{Data frame of metric memberships (long format, the same as the output of BCG.Membership).}

\item{df.rules}{Data frame of BCG model rules.}
}
\value{
Returns a data frame of results in the wide format.
}
\description{
Biological Condition Gradient Level assignment given metric memberships.
}
\details{
Input is metric memberships and a rules tables.  Output are
}
\examples{
library(readxl)

# Calculate Metrics
df.samps.bugs <- read_excel(system.file("./extdata/Data_BCG_PacNW.xlsx"
                                        , package="BCGcalc"))
myDF <- df.samps.bugs
df.metric.values.bugs <- metric.values(myDF, "bugs")

# Import Rules
df.rules <- read_excel(system.file("./extdata/Rules.xlsx"
                             , package="BCGcalc"), sheet="BCG_PacNW_2018") 

# Calculate Metric Memberships
df.Metric.Membership <- BCG.Metric.Membership(df.metric.values.bugs, df.rules)

# Calculate Level Memberships
df.Level.Membership <- BCG.Level.Membership(df.Metric.Membership, df.rules)

# Show results
View(df.Level.Membership)

# Save Results
write.table(df.Level.Membership, "Level.Membership.tsv"
            , row.names=FALSE, col.names=TRUE, sep="\\t")

}
