% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCG.Metric.Membership.R
\name{BCG.Metric.Membership}
\alias{BCG.Metric.Membership}
\title{BCG Metric Membership}
\usage{
BCG.Metric.Membership(df.metrics, df.rules, input.shape = "wide")
}
\arguments{
\item{df.metrics}{Wide data frame with metric values to be evaluated.}

\item{df.rules}{Data frame of metric thresholds to check.}

\item{input.shape}{Shape of df.metrics; wide or long.  Default is wide.}
}
\value{
Returns a data frame of results in the long format.
}
\description{
Biological Condition Gradient fuzzy membership for metrics.
}
\details{
Converts metric values into BCG membership values.  
Uses a rules table to define the metrics, scoring range, and direction for each named index.
}
\examples{
library(readxl)
library(BioMonTools)

# Calculate Metrics
df.samps.bugs <- read_excel(system.file("./extdata/Data_BCG_PacNW.xlsx"
                                        , package="BCGcalc")
                           , guess_max = 10^6)
myDF <- df.samps.bugs
df.metric.values.bugs <- metric.values(myDF, "bugs")

# Import Rules
df.rules <- read_excel(system.file("./extdata/Rules.xlsx"
                             , package="BCGcalc"), sheet="BCG_PacNW_v1_500ct") 

# Run function
df.Metric.Membership <- BCG.Metric.Membership(df.metric.values.bugs, df.rules)

# Show Results
View(df.Metric.Membership)

\dontrun{
# Save Results
write.table(df.Metric.Membership, "Metric.Membership.tsv"
              , row.names=FALSE, col.names=TRUE, sep="\\t")
}
}
