% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCGcalc.R
\docType{package}
\name{BCGcalc-package}
\alias{BCGcalc}
\alias{BCGcalc-package}
\title{BCGcalc: Metric calculation and other tasks related to BCG}
\description{
Enables users to employ various tools and calculations related 
to BCG.
}
\section{BCGcal functions}{

#' qc.checks
#' ...
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leppott/BCGcalc}
  \item Report bugs at \url{https://github.com/leppott/BCGcalc/issues}
}

}
\author{
\strong{Maintainer}: Erik W. Leppo \email{Erik.Leppo@tetratech.com}

Other contributors:
\itemize{
  \item Jen Stamp \email{Jen.Stamp@tetratech.com} [contributor]
  \item Matt Hedin \email{Matt.Hedin@tetratech.com} [contributor]
  \item Ben Block \email{Ben.Block@tetratech.com} [contributor]
}

}
