% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metrics_scoring}
\alias{metrics_scoring}
\title{Metric Scoring}
\format{
A data frame with 69 rows and 8 variables:
\describe{
   \item{Index.Name}{}
   \item{Index.Region}{}
   \item{Metric}{}
   \item{Direction}{}
   \item{Thresh.Lo}{}
   \item{Thresh.Hi}{}
   \item{ScoreRegime}{}
   \item{Name_Other}{}
}
}
\usage{
metrics_scoring
}
\description{
Long Description
}
\keyword{datasets}
