% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.checks.R
\name{qc.checks}
\alias{qc.checks}
\title{QC checks on metric values}
\usage{
qc.checks(df.metrics, df.checks, input.shape = "wide")
}
\arguments{
\item{df.metrics}{Wide data frame with metric values to be evaluated.}

\item{df.checks}{Data frame of metric thresholds to check.}

\item{input.shape}{Shape of df.metrics; wide or long.  Default is wide.}
}
\value{
Returns a data frame of SampleID checks and results; Pass and Fail.
}
\description{
Long Description
}
\details{
used reshape2 package
}
\examples{
library(readxl)

# Calculate Metrics
df.samps.bugs <- read_excel("./data-raw/Data_BCG_PacNW.xlsx")
myDF <- df.samps.bugs
df.metric.values.bugs <- metric.values(myDF, "bugs")

# Import Checks
df.checks.PacNW <- read_excel("./inst/extdata/MetricFlags.xlsx", sheet="Flags") 

# Run function
df.flags <- qc.checks(df.metric.values.bugs, df.checks.PacNW)

# show results
table(df.flags[,"CheckName"], df.flags[,"Flag"])

 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 junk below, need to remove
  #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  
# convert to long format
df.long <- reshape2::melt(df.metric.values.bugs, id.vars=c("SAMPLEID", "INDEX_NAME", "REGION")
                          , variable.name="metric.name", value.name="metric.value")


#' #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# don't need to cut down as merge will only keep the matches
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~#' 

# select for appropriate index
myIndex <- "BCG_PacNW_2018"
df.checks.PacNW <- df.checks[df.checks[,"Index_Name"]==myIndex, ]
# Metrics to Check
metrics.check <- as.vector(unique(as.data.frame(df.checks.PacNW[,"Metric"])))


#~~~~~~~~~~~~~~~~~~~~~~~~~
# select only metrics to check
boo.metrics <- df.long[, "metric.name"] \%in\% metrics.check
df.QC <- df.long[boo.metrics, ]
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

# temp, just ni_total 
df.QC <- df.long[df.long$metric.name=="ni_total", ]

# add QC checks
df.merge <- merge(df.long, df.checks, by.x=c("INDEX_NAME", "metric.name"), by.y=c("Index_Name", "Metric"))
# drops non-matching metrics



}
