% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxa_bugs_genus}
\alias{taxa_bugs_genus}
\title{USGS Gages}
\format{A data frame with 5066 rows and 20 variables:
\describe{
   \item{usgsGageID}{USGS Gage ID}
   \item{siteName}{USGS Site Name}
}}
\usage{
taxa_bugs_genus
}
\description{
Long Description
}
\keyword{datasets}
