% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxa_fish}
\alias{taxa_fish}
\title{USGS Gages}
\format{A data frame with 1693 rows and 15 variables:
\describe{
   \item{usgsGageID}{USGS Gage ID}
   \item{siteName}{USGS Site Name}
}}
\usage{
taxa_fish
}
\description{
Long Description
}
\keyword{datasets}
